Factored out facedrawing
