+++ src/Chart/common/drawFaces.css
@@ -1,0 +1,19 @@
+.faceActive0 {
+  fill: #008996;
+}
+
+.faceActive1 {
+  fill: #0099a9;
+}
+
+.faceActive2 {
+  fill: #35b8c4;
+}
+
+.faceActive3 {
+  fill: #9de2e8;
+}
+
+.faceInactive {
+  fill: #035d67;
+}
