+++ src/Chart/utils.js
@@ -29,6 +29,8 @@ exports.drawYAxisGroup = (svg, props) => {
         .attr("transform", `translate(0, ${props.margin.top})`);
 }
 
+exports.getChartGroup = (svg, styleClass) => svg.select(`.${styleClass}`);
+
 exports.drawChartGroup = (svg, props, styleClass) => {
   return svg.append('g')
         .attr('class', styleClass)
